% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_loglik.R
\name{cv_loglik}
\alias{cv_loglik}
\title{Performs cross-validation to calculate the average predicted log likelihood for the \code{logreg2ph} method. This function can be used to select the B-spline basis that yields the largest average predicted log likelihood.}
\usage{
cv_loglik(
  seed = 1,
  interp = TRUE,
  nfolds = 5,
  Y_unval = NULL,
  Y_val = NULL,
  X_unval = NULL,
  X_val = NULL,
  C = NULL,
  Validated = NULL,
  Bspline = NULL,
  data,
  theta_pred = NULL,
  gamma_pred = NULL,
  TOL = 1e-04,
  MAX_ITER = 1000
)
}
\arguments{
\item{seed}{(For reproducibility in assigning the folds) an integer to specify the random number generator.}

\item{interp}{Indicator of whether the B-spline coefficients in the testing data should be linearly interpolated from the training data. Defaults to \code{TRUE}.}

\item{nfolds}{Specifies the number of cross-validation folds. The default value is \code{5}. Although \code{nfolds} can be as large as the sample size (leave-one-out cross-validation), it is not recommended for large datasets. The smallest value allowable is \code{3}.}

\item{Y_unval}{Column name with the unvalidated outcome. If \code{Y_unval} is null, the outcome is assumed to be error-free.}

\item{Y_val}{Column name with the validated outcome.}

\item{X_unval}{Column name(s) with the unvalidated predictors.  If \code{X_unval} and \code{X_val} are \code{null}, all precictors are assumed to be error-free.}

\item{X_val}{Column name(s) with the validated predictors. If \code{X_unval} and \code{X_val} are \code{null}, all precictors are assumed to be error-free.}

\item{C}{(Optional) Column name(s) with additional error-free covariates.}

\item{Validated}{Column name with the validation indicator. The validation indicator can be defined as \code{Validated = 1} or \code{TRUE} if the subject was validated and \code{Validated = 0} or \code{FALSE} otherwise.}

\item{Bspline}{Vector of column names containing the B-spline basis functions.}

\item{data}{A dataframe with one row per subject containing columns: \code{Y_unval}, \code{Y_val}, \code{X_unval}, \code{X_val}, \code{C}, \code{Validated}, and \code{Bspline}.}

\item{theta_pred}{Vector of columns in \code{data} that pertain to the predictors in the analysis model.}

\item{gamma_pred}{Vector of columns in \code{data} that pertain to the predictors in the outcome error model.}

\item{TOL}{Tolerance between iterations in the EM algorithm used to define convergence. Defaults to \code{1E-4}.}

\item{MAX_ITER}{Maximum number of iterations allowed in the EM algorithm. Defaults to \code{1000}.}
}
\value{
\item{avg_pred_loglike}{Stores the average predicted log likelihood.}
\item{pred_loglike}{Stores the predicted log likelihoood in each fold.}
\item{converged}{Stores the convergence status of the EM algorithm in each run.}
}
\description{
Performs cross-validation to calculate the average predicted log likelihood for the \code{logreg2ph} method. This function can be used to select the B-spline basis that yields the largest average predicted log likelihood.
}
